+++ src/styles/globalStyles.tsx
